Finished part4

+++ part4/bloglist/controllers/blogs.js
@@ -1,8 +1,12 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+
 
 blogsRouter.get('/', async (request, response) => {
-    const blogs = await Blog.find({})
+    const blogs = await Blog
+        .find({}).populate('user', {username: 1, name: 1})
+
     response.json(blogs)
 })
 
@@ -20,15 +24,33 @@ blogsRouter.post('/', async (request, response) => {
         return response.status(400).end()
     }
 
+    if(!request.user){
+        return response.status(401).json({error: 'must be logged in'})
+    } 
+
+    const user = await User.findById(request.user.id)
+    if (!user) return response.status(401).json({ error: 'must be logged in' })
+
     const blog = new Blog({
         ...request.body,
         likes: request.body.likes ?? 0,
+        user: user.id
     })
-    const result = await blog.save()
-    response.status(201).json(result)
+    const savedBlog = await blog.save()
+
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
+
+    response.status(201).json(savedBlog)
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+    if(!request.user) return response.status(401).json({error: 'must be logged in'})
+
+    const blog = await Blog.findById(request.params.id)
+
+    if(blog?.user?.toString() !== request.user.id) return response.status(401).json({ error: 'you can not delete others blogs' })
+
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
 })
